requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
umap-learn>=0.5.3
tqdm>=4.65.0
numpy>=1.24.0
seaborn>=0.12.0
datasets>=2.14.0
